Item animations now rotate along x-rotation
