+++ src/main/resources/assets/industrialwarfare/animations/third_person/vetterli_t.animation.json
@@ -60,7 +60,7 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_hidden bullet true; lock_limbs true"
+				"0.0": "set_hidden cartridge true"
 			}
 		},
 		"ads_aiming_start": {
@@ -144,7 +144,7 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_hidden bullet true; lock_limbs true"
+				"0.0": "set_hidden cartridge true"
 			}
 		},
 		"ads_aiming_stop": {
@@ -235,8 +235,7 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_bullet hidden true; lock_limbs true",
-				"0.5": "lock_limbs false"
+				"0.0": "set_hidden cartridge true"
 			}
 		},
 		"ads_firing": {
@@ -317,7 +316,7 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_bullet hidden true; lock_limbs true; move firing_pin 0 0 0"
+				"0.0": "set_hidden cartridge true; move firing_pin 0 0 0"
 			}
 		},
 		"hip_aiming": {
@@ -389,7 +388,7 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_hidden bullet true; lock_limbs false"
+				"0.0": "set_hidden cartridge true; lock_limbs false"
 			}
 		},
 		"hip_firing": {
@@ -495,12 +494,12 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_hidden bullet true"
+				"0.0": "set_hidden cartridge true"
 			}
 		},
 		"hip_cycling": {
 			"loop": "hold_on_last_frame",
-			"animation_length": 1.5,
+			"animation_length": 1,
 			"bones": {
 				"firearm": {
 					"rotation": {
@@ -580,7 +579,7 @@
 						}
 					}
 				},
-				"bullet": {
+				"cartridge": {
 					"rotation": {
 						"vector": [0, -7.5, 0]
 					},
@@ -641,9 +640,6 @@
 						},
 						"0.9583": {
 							"vector": [-31.33344, -6.68018, -12.0467]
-						},
-						"1.5": {
-							"vector": [-31.33344, -6.68018, -12.0467]
 						}
 					},
 					"position": {
@@ -672,9 +668,6 @@
 						},
 						"0.9583": {
 							"vector": [0, -0.5, 1.5]
-						},
-						"1.5": {
-							"vector": [0, -0.5, 1.5]
 						}
 					}
 				}
@@ -688,9 +681,9 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_hidden bullet false; move firing_pin 0 0 0",
+				"0.0": "set_hidden cartridge false; move firing_pin 0 0 0",
 				"0.375": "move firing_pin 0 0 0.25",
-				"1.0": "set_hidden bullet true"
+				"1.0": "set_hidden cartridge true"
 			}
 		},
 		"port_arms": {
@@ -770,7 +763,7 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_hidden bullet true"
+				"0.0": "set_hidden cartridge true"
 			}
 		},
 		"reload_start": {
@@ -846,10 +839,21 @@
 					"position": {
 						"vector": [0, -0.5, 1.5]
 					}
+				},
+				"head": {
+					"rotation": {
+						"0.0": {
+							"vector": [0, 0, 0]
+						},
+						"0.25": {
+							"vector": [45, 32.5, 0],
+							"easing": "easeInOutSine"
+						}
+					}
 				}
 			},
 			"timeline": {
-				"0.0": "set_hidden bullet true"
+				"0.0": "set_hidden cartridge true; lock_limbs true"
 			}
 		},
 		"reload": {
@@ -865,7 +869,7 @@
 						"easing": "easeInOutSine"
 					}
 				},
-				"bullet": {
+				"cartridge": {
 					"rotation": {
 						"0.5": {
 							"vector": [-4.83711, -5.73853, 40.24258],
@@ -939,6 +943,12 @@
 					"position": {
 						"vector": [0, 0, 0]
 					}
+				},
+				"head": {
+					"rotation": {
+						"vector": [45, 32.5, 0],
+						"easing": "easeInOutSine"
+					}
 				}
 			},
 			"sound_effects": {
@@ -947,13 +957,13 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_hidden bullet false",
-				"0.6667": "set_hidden bullet true"
+				"0.0": "set_hidden cartridge false; lock_limbs true",
+				"0.6667": "set_hidden cartridge true"
 			}
 		},
 		"reload_end": {
 			"loop": "hold_on_last_frame",
-			"animation_length": 2,
+			"animation_length": 1.5,
 			"bones": {
 				"firearm": {
 					"rotation": {
@@ -1110,9 +1120,21 @@
 						}
 					}
 				},
-				"bullet": {
+				"cartridge": {
 					"position": {
 						"vector": [-2, -3, 0]
+					}
+				},
+				"head": {
+					"rotation": {
+						"0.0": {
+							"vector": [45, 32.5, 0],
+							"easing": "easeInOutSine"
+						},
+						"0.25": {
+							"vector": [0, 0, 0],
+							"easing": "easeInOutSine"
+						}
 					}
 				}
 			},
@@ -1127,12 +1149,12 @@
 			"timeline": {
 				"0.0": "set_hidden bullet true; lock_limbs false",
 				"0.9167": "move firing_pin 0 0 0.25",
-				"2.0": "set_hidden bullet true"
+				"1.5": "set_hidden cartridge true; lock_limbs false"
 			}
 		},
 		"reload_end_extract": {
 			"loop": "hold_on_last_frame",
-			"animation_length": 2,
+			"animation_length": 1.5,
 			"bones": {
 				"firearm": {
 					"rotation": {
@@ -1198,7 +1220,7 @@
 						}
 					}
 				},
-				"bullet": {
+				"cartridge": {
 					"rotation": {
 						"vector": [0, -7.5, 0]
 					},
@@ -1297,6 +1319,18 @@
 							"vector": [0, 0, 0]
 						}
 					}
+				},
+				"head": {
+					"rotation": {
+						"0.0": {
+							"vector": [45, 32.5, 0],
+							"easing": "easeInOutSine"
+						},
+						"0.25": {
+							"vector": [0, 0, 0],
+							"easing": "easeInOutSine"
+						}
+					}
 				}
 			},
 			"sound_effects": {
@@ -1308,10 +1342,10 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_hidden bullet true; lock_limbs false",
-				"0.75": "set_hidden bullet false",
+				"0.0": "set_hidden cartridge true; lock_limbs true",
+				"0.75": "set_hidden cartridge false",
 				"0.9167": "move firing_pin 0 0 0.25",
-				"1.2083": "set_hidden bullet true"
+				"1.5": "set_hidden cartridge true; lock_limbs false"
 			}
 		},
 		"reload_hold": {
@@ -1340,7 +1374,7 @@
 						}
 					}
 				},
-				"bullet": {
+				"cartridge": {
 					"position": {
 						"0.0": {
 							"vector": [-2, -3, 0]
@@ -1390,7 +1424,7 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_hidden bullet true"
+				"0.0": "set_hidden cartridge true"
 			}
 		},
 		"trail_arms": {
@@ -1419,7 +1453,7 @@
 				}
 			},
 			"timeline": {
-				"0.0": "set_hidden bullet true"
+				"0.0": "set_hidden cartridge true"
 			}
 		}
 	}
